+++ src/Symfony/Component/Security/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 * Removed `Argon2iPasswordEncoder`, use `SodiumPasswordEncoder` instead
 * Removed `BcryptPasswordEncoder`, use `NativePasswordEncoder` instead
+* Removed the `has_role()` function from security expressions, use `is_granted()` instead.
 
 4.3.0
 -----